refactor(cli): converge ref-merge importers on the shared module
updateRefs and the doctor package check import mergeAndParseRefs from
the store-free module that #682 landed, instead of reaching through
runtime.ts and its bootStore/ke transitive imports.
